requirements.txt
anthropic
python-dotenv
openai
fastapi
uvicorn
pydantic